refactor(domain/dto): altera o local do DTO do create-user.dto e altera a nomenclatura
