+++ src/shared/injection/container.ts
@@ -4,7 +4,7 @@ import { Database } from "@infrastructure/database/connection";
 import { TYPES } from "@shared/constants/constants";
 import { patientsModule } from "@infrastructure/injection/patients.module.container";
 import { authModule } from "@infrastructure/injection/auth.module.container";
-import { usersModule } from "@infrastructure/injection/user.module.container";
+import { usersModule } from "@infrastructure/injection/users.module.container";
 
 const container = new Container();
