+++ tests/mikro-orm.test.config.ts
@@ -1,34 +1,32 @@
-import { MariaDbDriver, type Options } from "@mikro-orm/mariadb";
-import { TsMorphMetadataProvider } from "@mikro-orm/reflection";
-import { configDotenv } from "dotenv";
+    import { MariaDbDriver, type Options } from "@mikro-orm/mariadb";
+    import { TsMorphMetadataProvider } from "@mikro-orm/reflection";
+    import { configDotenv } from "dotenv";
 
-configDotenv();
+    configDotenv();
 
-const port = parseInt(process.env.TEST_DB_PORT || "3306");
+    const config: Options = {
+        driver: MariaDbDriver,
+        dbName: process.env.TEST_DB_NAME,
+        entities: ["src/domain/entities/*.entity.ts"],
+        metadataProvider: TsMorphMetadataProvider,
 
-const config: Options = {
-    driver: MariaDbDriver,
-    dbName: process.env.TEST_DB_NAME,
-    entities: ["src/domain/entities/*.entity.ts"],
-    metadataProvider: TsMorphMetadataProvider,
-
-    debug: false,
-    driverOptions: {
-        connection: {
-            connectTimeout: 10000,
-            socketTimeout: 30000,
+        debug: false,
+        driverOptions: {
+            connection: {
+                connectTimeout: 10000,
+                socketTimeout: 30000,
+            },
         },
-    },
 
-    migrations: {
-        transactional: false,
-        path: "../src/infrastructure/database/migrations",
-    },
+        migrations: {
+            transactional: false,
+            path: "../src/infrastructure/database/migrations",
+        },
 
-    user: process.env.TEST_DB_USER,
-    host: process.env.TEST_DB_HOST,
-    port: port,
-    password: process.env.TEST_DB_PASSWORD,
-};
+        user: process.env.TEST_DB_USER,
+        host: process.env.TEST_DB_HOST,
+        port: 3306,
+        password: process.env.MARIADB_ROOT_PASSWORD_TEST,
+    };
 
-export default config;
+    export default config;
